empesada la dal

+++ BuisnessLayer/interfaces/IBL_Admin.cs
@@ -25,7 +25,7 @@ namespace BuisnessLayer.interfaces
         EVehiculo crearVehiculos(string Marca, string Modelo, string Matrícula, int cantAsientos);
         EVehiculo editarVehiculos(string Marca, string Modelo, string Matrícula, int cantAsientos);
 
-        EParada crearParada(string nombre, float lat, float lon);
+        EParada crearParada(string nombre, double lat, double lon);
 
         ELinea crearLinea(string nombre);
